backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
biopython==1.81 